% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assingments.R
\name{process_assignments}
\alias{process_assignments}
\title{Process assignments create with edu and completed by students}
\usage{
process_assignments(path, outfile)
}
\arguments{
\item{outfile}{Name of the the output}

\item{folder}{Absolute path to the folder containing the completed assignments as R.scripts ending either with .R or .r}
}
\value{
A R-script
}
\description{
Function processes number of completed assignments and outputs a single R-script with the question and all students answers AND correct answeers
}
\examples{
 \dontrun{
 process_assignments(path = "~/rcourse/exercises/import/", outfile = "import_summary")
 }

}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
