% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assingments.R
\name{process_assignments}
\alias{process_assignments}
\title{Process assignments create with edu and completed by students}
\usage{

  process_assignments(path_assignments = "~/btsync/workspace/ropengov/edutest/basics",
  outfile_name = "basics_summary",
  outfile_path = "~/btsync/workspace/ropengov/edutest/")
}
\arguments{
\item{path_assignments}{Absolute path to the folder containing the completed assignments as R.scripts ending either with .R or .r}

\item{outfile_name}{Name of the the output script}

\item{outfile_path}{Absolute path to the folder where scipt should be written}
}
\value{
A R-script
}
\description{
Function processes number of completed assignments and outputs a single R-script with the question and all students answers AND correct answeers
}
\examples{
 \dontrun{
 process_assignments(path_assignments = "~/btsync/workspace/ropengov/edutest/basics",
                   outfile_name = "basics_summary",
                   outfile_path = "~/btsync/workspace/ropengov/edutest/")
 }

}
\author{
Markus Kainu <markus.kainu@kela.fi>
}
